move some structs
